refactor(engineer_loop/selection): split into mod+issue_extraction+sanitization+selectors_extra (#1266)
selection.rs 758 -> mod.rs 398 + issue_extraction.rs 148 + sanitization.rs 83 + selectors_extra.rs 169.
issue_extraction.rs: extract_existing_issue_number + helpers.
sanitization.rs: strip_shell_unsafe, sanitize_issue_title, sanitize_commit_message.
selectors_extra.rs: select_open_issue, select_cargo_action, is_*_achievable.
